pyproject.toml
[tool.poetry]
name = "pybase"
version = "0.1.0"
description = ""
authors = ["Jose Fco de la Fuente <josefcodelafuente@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.6.1"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-sugar = "^0.9.7"
pytest-cov = "^4.1.0"
pre-commit = "^3.5.0"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
